refactor(fretboard): extract note data model
Move note classification into utils/semantics.ts and extract the
per-note computation loop into the useNoteData hook, reducing
FretboardSVG.tsx to pure rendering concerns.
