+++ src/components/FretboardSVG/FretboardSVG.tsx
@@ -1,21 +1,17 @@
 import { useId, useMemo, useCallback, memo, useState, useEffect, useRef, type CSSProperties, type ReactNode } from "react";
 import { clsx } from "clsx";
 import {
-  NOTES,
-  ENHARMONICS,
   getNoteDisplay,
-  getNoteDisplayInScale,
-  INTERVAL_NAMES,
   formatAccidental,
-  SCALES,
   type PracticeLens,
   type NoteSemantics,
 } from "../../core/theory";
-import { parseNote } from "../../core/guitar";
 import { STRING_ROW_PX_TABLET } from "../../layout/responsive";
 import styles from "./FretboardSVG.module.css";
 import { useFretboardGeometry } from "./hooks/useFretboardGeometry";
-import type { ShapePolygon, CagedShape } from "../../shapes";
+import { useNoteData } from "./hooks/useNoteData";
+import { getNoteVisuals, type BoxBound } from "./utils/semantics";
+import type { ShapePolygon } from "../../shapes";
 import type { ActiveShapeType } from "../../hooks/useFretboardState";
 import {
   NECK_BORDER,
@@ -27,12 +23,6 @@ import {
   NOTE_FONT_RATIO,
   INLAY_RADIUS_RATIO,
   INLAY_RADIUS_MIN,
-  RADIUS_SCALE_KEY_TONIC,
-  RADIUS_SCALE_CHORD_ROOT,
-  RADIUS_SCALE_CHORD_TONE,
-  RADIUS_SCALE_NOTE_ACTIVE,
-  RADIUS_SCALE_COLOR_TONE,
-  RADIUS_SCALE_DEFAULT,
 } from "../../core/constants";
 
 interface FretboardSVGProps {
@@ -47,7 +37,7 @@ interface FretboardSVGProps {
   highlightNotes: string[];
   rootNote: string;
   displayFormat?: "notes" | "degrees" | "none";
-  boxBounds?: { minFret: number; maxFret: number }[];
+  boxBounds?: BoxBound[];
   chordTones?: string[];
   chordRoot?: string;
   chordFretSpread?: number;
@@ -91,169 +81,7 @@ interface FretboardSVGProps {
   ) => void;
 }
 
-type BoxBound = { minFret: number; maxFret: number };
 
-type LensEmphasis = {
-  glowColor?: "cyan" | "orange" | "violet";
-  radiusBoost: number;
-  opacityBoost: number;
-};
-
-function getLensEmphasis(
-  noteClass: string,
-  practiceLens: PracticeLens | undefined,
-  isGuideTone: boolean,
-  isTension: boolean,
-): LensEmphasis {
-  const defaultEmphasis: LensEmphasis = { radiusBoost: 1, opacityBoost: 1 };
-
-  if (!practiceLens) return defaultEmphasis;
-
-  switch (practiceLens) {
-    case "guide-tones":
-      if (isGuideTone) {
-        return { glowColor: "cyan", radiusBoost: 1.15, opacityBoost: 1 };
-      }
-      if (noteClass.includes("chord-") || noteClass.includes("color-")) {
-        return { radiusBoost: 0.85, opacityBoost: 0.7 };
-      }
-      return defaultEmphasis;
-
-    case "tension":
-      if (isTension) {
-        return { glowColor: "orange", radiusBoost: 1.15, opacityBoost: 1 };
-      }
-      if (noteClass.includes("chord-")) {
-        return { radiusBoost: 0.85, opacityBoost: 0.7 };
-      }
-      return defaultEmphasis;
-
-    case "targets":
-    default:
-      return defaultEmphasis;
-  }
-}
-
-function classifyNote(
-  isScaleRoot: boolean,
-  isChordRootNote: boolean,
-  isColorNote: boolean,
-  isHighlighted: boolean,
-  isChordTone: boolean,
-  hasChordOverlay: boolean,
-  isChordInRange: boolean,
-  isInActiveShape: boolean,
-  shapePolygons: ShapePolygon[],
-  boxBounds: BoxBound[],
-  fretIndex: number,
-): string {
-  if (!hasChordOverlay) {
-    if (isScaleRoot && isHighlighted) return "key-tonic";
-    if (isColorNote && isHighlighted) return "note-blue";
-    if (isHighlighted) return "note-active";
-    if (
-      isColorNote &&
-      shapePolygons.length > 0 &&
-      boxBounds.some(
-        (b) => fretIndex >= b.minFret - 1 && fretIndex <= b.maxFret + 1,
-      )
-    )
-      return "note-blue";
-    return "note-inactive";
-  }
-
-  if (isChordRootNote && isChordTone && isChordInRange && isInActiveShape) return "chord-root";
-  if (isHighlighted && isChordTone && isChordInRange && isInActiveShape) return "chord-tone-in-scale";
-  if (isHighlighted && isColorNote && isInActiveShape) return "color-tone";
-  if (isHighlighted && isInActiveShape) return "scale-only";
-  if (!isHighlighted && isChordTone && isChordInRange && isInActiveShape)
-    return "chord-tone-outside-scale";
-  return "note-inactive";
-}
-
-function classifyNoteFromSemantics(
-  sem: NoteSemantics,
-  isChordInRange: boolean,
-  isInActiveShape: boolean,
-  hasChordOverlay: boolean,
-  isHighlighted: boolean,
-  shapePolygons: ShapePolygon[],
-  boxBounds: BoxBound[],
-  fretIndex: number,
-): string {
-  if (!hasChordOverlay) {
-    return classifyNote(
-      sem.isScaleRoot, sem.isChordRoot, sem.isColorTone, isHighlighted,
-      sem.isChordTone, hasChordOverlay, isChordInRange, isInActiveShape, shapePolygons, boxBounds, fretIndex,
-    );
-  }
-
-  if (sem.isChordRoot && sem.isChordTone && isChordInRange && isInActiveShape) return "chord-root";
-  if (sem.isInScale && sem.isChordTone && isChordInRange && isInActiveShape) return "chord-tone-in-scale";
-  if (sem.isInScale && sem.isColorTone && isInActiveShape && isHighlighted) return "color-tone";
-  if (sem.isInScale && isInActiveShape && isHighlighted) return "scale-only";
-  if (sem.isChordTone && isChordInRange && isInActiveShape) return "chord-tone-outside-scale";
-  return "note-inactive";
-}
-
-type NoteShape = "circle" | "squircle" | "diamond" | "hexagon";
-
-type NoteVisuals = {
-  radiusScale: number;
-  noteShape: NoteShape;
-};
-
-function getNoteVisuals(
-  noteClass: string,
-): NoteVisuals {
-  switch (noteClass) {
-    case "key-tonic":
-      return {
-        radiusScale: RADIUS_SCALE_KEY_TONIC,
-        noteShape: "circle",
-      };
-    case "chord-root":
-      return {
-        radiusScale: RADIUS_SCALE_CHORD_ROOT,
-        noteShape: "squircle",
-      };
-    case "chord-tone-in-scale":
-      return {
-        radiusScale: RADIUS_SCALE_CHORD_TONE,
-        noteShape: "squircle",
-      };
-    case "note-active":
-      return {
-        radiusScale: RADIUS_SCALE_NOTE_ACTIVE,
-        noteShape: "circle",
-      };
-    case "note-blue":
-      return {
-        radiusScale: RADIUS_SCALE_NOTE_ACTIVE,
-        noteShape: "hexagon",
-      };
-    case "scale-only":
-      return {
-        radiusScale: RADIUS_SCALE_NOTE_ACTIVE,
-        noteShape: "circle",
-      };
-    case "color-tone":
-      return {
-        radiusScale: RADIUS_SCALE_COLOR_TONE,
-        noteShape: "hexagon",
-      };
-    case "chord-tone-outside-scale":
-      return {
-        radiusScale: RADIUS_SCALE_CHORD_TONE,
-        noteShape: "diamond",
-      };
-    default:
-      return {
-        radiusScale: RADIUS_SCALE_DEFAULT,
-        noteShape: "circle",
-      };
-  }
-}
 
 /** Pre-rasterizes wood-grain layers to PNG for GPU-composited scrolling. */
 function useWoodGrainTexture(width: number, height: number): string | null {
@@ -630,216 +458,33 @@ export const FretboardSVG = memo(function FretboardSVG({
     });
   }, [totalColumns, startFret, stringRowPx, svgDefUrl, fretCenterX, inlayYAt, inlayYBottomAt, inlayYTopAt]);
 
-  const noteData = useMemo(() => {
-    const notes = [];
-    const scale = SCALES[scaleName] || [];
-    const normRoot = rootNote && (ENHARMONICS[rootNote]?.includes("b") ? ENHARMONICS[rootNote] : rootNote);
-    const rootIdx = rootNote ? NOTES.indexOf(normRoot.includes("#") ? normRoot : rootNote) : -1;
-
-    // Pre-calculate normalized hidden notes for faster lookup
-    const normalizedHidden = new Set<string>();
-    if (hiddenNotes && hiddenNotes.size > 0) {
-      hiddenNotes.forEach(n => {
-        normalizedHidden.add(n);
-        // Also hide enharmonic equivalents if it's a note name
-        const enh = ENHARMONICS[n];
-        if (enh) normalizedHidden.add(enh);
-      });
-    }
-
-    // Pre-calculate highlighted notes set for faster lookup
-    const highlightSet = new Set(highlightNotes);
-
-    // Pre-calculate chord tones set
-    const chordToneSet = new Set(chordTones);
-
-    // Pre-calculate color notes set
-    const colorNoteSet = new Set(colorNotes);
-
-    for (let stringIndex = 0; stringIndex < numStrings; stringIndex++) {
-      const layoutRow = fretboardLayout[stringIndex];
-      const openNoteName = parseNote(tuning[stringIndex])?.noteName;
-
-      for (let idx = 0; idx <= totalColumns; idx++) {
-        const fretIndex = startFret + idx;
-        if (fretIndex >= maxFret) continue;
-
-        const noteName = layoutRow[fretIndex];
-
-        const isNoteHidden = normalizedHidden.has(noteName) || normalizedHidden.has(`${stringIndex}-${fretIndex}`);
-
-        const isHighlighted =
-          !isNoteHidden &&
-          (highlightSet.has(noteName) ||
-            highlightSet.has(`${stringIndex}-${fretIndex}`));
-        
-        const isChordTone =
-          !isNoteHidden && hasChordOverlay && chordToneSet.has(noteName);
-        
-        const isScaleRoot =
-          !isNoteHidden &&
-          (noteName === rootNote ||
-            ENHARMONICS[noteName] === rootNote ||
-            ENHARMONICS[rootNote] === noteName);
-        
-        const isChordRootNote =
-          !isNoteHidden &&
-          !!chordRoot &&
-          (noteName === chordRoot ||
-            ENHARMONICS[noteName] === chordRoot ||
-            ENHARMONICS[chordRoot] === noteName);
-        
-        const isColorNote = !!(!isNoteHidden && colorNoteSet.size > 0 && (
-          colorNoteSet.has(noteName) || 
-          (ENHARMONICS[noteName] && colorNoteSet.has(ENHARMONICS[noteName]!))
-        ));
-
-        const isInsideAnyPolygon = shapePolygons.some((poly) => {
-          const leftFret = poly.vertices[stringIndex]?.fret;
-          const rightFret =
-            poly.vertices[poly.vertices.length - 1 - stringIndex]?.fret;
-          return (
-            leftFret !== undefined &&
-            rightFret !== undefined &&
-            fretIndex >= leftFret &&
-            fretIndex <= rightFret
-          );
-        });
-
-        // Shape-aware, spread-aware playable context.
-        // True when this note coordinate should receive chord overlay emphasis.
-        // Checks only the active shape(s) — not all polygons — so the spread
-        // buffer correctly extends the active shape boundary, not any visible shape.
-        const isInPlayableContext: boolean = (() => {
-          if (!hasChordOverlay) return false;
-          // 3NPS has no polygon shapes; gate chord overlay by aggregate fret bounds
-          if (activePattern === "3nps" && shapeScope !== "global" && boxBounds.length > 0) {
-            return boxBounds.some(
-              (b) =>
-                fretIndex >= b.minFret - chordFretSpread &&
-                fretIndex <= b.maxFret + chordFretSpread,
-            );
-          }
-          if (shapePolygons.length === 0 || !activePattern) return true;
-          if (shapeScope === "global") return true;
-          return shapePolygons.some((poly) => {
-            if (shapeScope === "single") {
-              if (activePattern === "caged" && poly.shape !== activeShape) return false;
-              if (activePattern === "3nps" && poly.shape !== activeShape) return false;
-            } else if (shapeScope === "multi" && Array.isArray(activeShape)) {
-              if (!(activeShape as CagedShape[]).includes(poly.shape as CagedShape)) return false;
-            }
-            const leftFret = poly.vertices[stringIndex]?.fret;
-            const rightFret = poly.vertices[poly.vertices.length - 1 - stringIndex]?.fret;
-            if (leftFret === undefined || rightFret === undefined) return false;
-            const clampedLeft = Math.min(maxFret, Math.max(0, leftFret));
-            const clampedRight = Math.min(maxFret, Math.max(0, rightFret));
-            if (clampedLeft > clampedRight) return false;
-            return (
-              fretIndex >= clampedLeft - chordFretSpread &&
-              fretIndex <= clampedRight + chordFretSpread
-            );
-          });
-        })();
-
-        // Both range and shape-membership now derive from the same computation.
-        const isChordInRange = isInPlayableContext;
-        const isInActiveShape = isInPlayableContext || !hasChordOverlay || !activePattern;
-
-        // Use the composable semantic model when available; fall back to booleans.
-        // Keys are sharp-normalized (per project convention) so no enharmonic lookup needed.
-        const semantics = noteSemantics?.get(noteName);
-
-        const noteClass = isNoteHidden
-          ? "note-inactive"
-          : semantics
-            ? classifyNoteFromSemantics(
-                semantics,
-                isChordInRange,
-                isInActiveShape,
-                hasChordOverlay,
-                isHighlighted,
-                shapePolygons,
-                boxBounds,
-                fretIndex,
-              )
-            : classifyNote(
-                isScaleRoot,
-                isChordRootNote,
-                isColorNote,
-                isHighlighted,
-                isChordTone,
-                hasChordOverlay,
-                isChordInRange,
-                isInActiveShape,
-                shapePolygons,
-                boxBounds,
-                fretIndex,
-              );
-
-        if (noteClass === "note-inactive") continue;
-
-        let displayValue = getNoteDisplayInScale(
-          noteName,
-          rootNote,
-          scale,
-          useFlats,
-        );
-        if (displayFormat === "degrees" && rootNote) {
-          const noteIdx = NOTES.indexOf(noteName);
-          if (rootIdx !== -1 && noteIdx !== -1) {
-            displayValue = INTERVAL_NAMES[(noteIdx - rootIdx + 12) % 12];
-          }
-        } else if (
-          fretIndex === 0 &&
-          openNoteName === noteName
-        ) {
-          displayValue = getNoteDisplayInScale(
-            noteName,
-            rootNote,
-            scale,
-            useFlats,
-          );
-        }
-
-        const isWrapped = wrappedNotes.has(`${stringIndex}-${fretIndex}`);
-        const applyDimOpacity =
-          (shapePolygons.length > 0 &&
-            !isInsideAnyPolygon &&
-            (noteClass === "note-blue" ||
-              noteClass === "chord-tone-outside-scale" ||
-              noteClass === "chord-tone-in-scale" ||
-              noteClass === "chord-root" ||
-              noteClass === "key-tonic")) ||
-          (isWrapped && isHighlighted);
-
-        // Lens emphasis only applies when chord overlay is active.
-        // Without an overlay, no lens-driven dimming or emphasis should alter scale notes.
-        const lensEmphasis = getLensEmphasis(
-          noteClass,
-          hasChordOverlay ? practiceLens : undefined,
-          semantics?.isGuideTone ?? false,
-          semantics?.isTension ?? false,
-        );
-
-        const isHidden = false;
-
-        notes.push({
-          stringIndex,
-          fretIndex,
-          noteName,
-          noteClass,
-          displayValue,
-          applyDimOpacity,
-          applyLensEmphasis: lensEmphasis,
-          isHidden,
-          isTension: semantics?.isTension ?? false,
-          isGuideTone: semantics?.isGuideTone ?? false,
-        });
-      }
-    }
-    return notes;
-  }, [numStrings, fretboardLayout, totalColumns, startFret, maxFret, hiddenNotes, highlightNotes, hasChordOverlay, chordTones, rootNote, chordRoot, colorNotes, shapePolygons, boxBounds, chordFretSpread, scaleName, useFlats, displayFormat, wrappedNotes, practiceLens, tuning, noteSemantics, activePattern, activeShape, shapeScope]);
+  const noteData = useNoteData({
+    numStrings,
+    fretboardLayout,
+    totalColumns,
+    startFret,
+    maxFret,
+    hiddenNotes,
+    highlightNotes,
+    hasChordOverlay,
+    chordTones,
+    rootNote,
+    chordRoot,
+    colorNotes,
+    shapePolygons,
+    boxBounds,
+    chordFretSpread,
+    activePattern,
+    shapeScope,
+    activeShape,
+    scaleName: scaleName || "",
+    useFlats,
+    displayFormat,
+    wrappedNotes,
+    practiceLens,
+    tuning,
+    noteSemantics,
+  });
 
   return (
     <div
